# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "该服务用以生成短链接",
        "title": "短链接服务接口文档",
        "contact": {
            "name": "lomogo"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/email": {
            "post": {
                "description": "将生成的二维码和短链接发送到用户邮箱，链接必须以：http://或https://开头。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "邮件"
                ],
                "summary": "发送邮件",
                "parameters": [
                    {
                        "description": "需要转换的网址",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EmailInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": \"0\", \"msg\": \"成功\", \"details\": \"\", data:\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/link": {
            "post": {
                "description": "将用户传入的网址转换为短链接, 链接必须以：http://或https://开头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链"
                ],
                "summary": "生成短链",
                "parameters": [
                    {
                        "description": "需要转换的网址",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.URLInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": \"0\", \"msg\": \"成功\", \"details\": \"\", data:\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/qrcode": {
            "post": {
                "description": "将用户传入的网址转换为二维码，链接必须以：http://或https://开头",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "二维码"
                ],
                "summary": "生成二维码",
                "parameters": [
                    {
                        "description": "需要转换的网址",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.URLInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": \"0\", \"msg\": \"成功\", \"details\": \"\", data:\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.EmailInfo": {
            "type": "object",
            "properties": {
                "recipient": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "request.URLInfo": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}